# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Atualizado em 21/03/2021 02:47",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/search": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Buscar Usuários",
                "description": "Buscar Usuários",
                "operationId": "user-search",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de busca",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Quantidade de resultados por página",
                        "required": false
                    },
                    {
                        "name": "orderby",
                        "in": "query",
                        "description": "Nome da coluna de referência para ordenação",
                        "required": false
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordem asc ou desc",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/user/find/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Retornar Usuário por id",
                "description": "Retornar Usuário por id",
                "operationId": "user-find",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de order",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados de order"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/user/save": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Salvar Usuários",
                "description": "Salvar Usuários",
                "operationId": "user-save",
                "responses": {
                    "200": {
                        "description": "Dados salvos"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/user/delete/{id}": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Deletar Usuário",
                "description": "Deletar Usuário",
                "operationId": "user-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de order",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados salvos"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/user/types": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Listar opções de type",
                "description": "Listar opções de type",
                "operationId": "user-types",
                "responses": {
                    "200": {
                        "description": "Dados"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/user/type/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Pegar opção de type por ID",
                "description": "Pegar opção de type por ID",
                "operationId": "user-type",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de order",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/user-notification/search": {
            "get": {
                "tags": [
                    "UserNotification"
                ],
                "summary": "Buscar User_notifications",
                "description": "Buscar User_notifications",
                "operationId": "user-notification-search",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de busca",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Quantidade de resultados por página",
                        "required": false
                    },
                    {
                        "name": "orderby",
                        "in": "query",
                        "description": "Nome da coluna de referência para ordenação",
                        "required": false
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordem asc ou desc",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/user-notification/find/{id}": {
            "get": {
                "tags": [
                    "UserNotification"
                ],
                "summary": "Retornar User_notification por id",
                "description": "Retornar User_notification por id",
                "operationId": "user-notification-find",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de order",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados de order"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/user-notification/save": {
            "post": {
                "tags": [
                    "UserNotification"
                ],
                "summary": "Salvar User_notifications",
                "description": "Salvar User_notifications",
                "operationId": "user-notification-save",
                "responses": {
                    "200": {
                        "description": "Dados salvos"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        },
        "/api/user-notification/delete/{id}": {
            "post": {
                "tags": [
                    "UserNotification"
                ],
                "summary": "Deletar User_notification",
                "description": "Deletar User_notification",
                "operationId": "user-notification-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de order",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados salvos"
                    },
                    "500": {
                        "description": "{'message':'Erro', 'fields':[]}"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}